Reverting to using long for ID and adding migration

+++ FitnessCelebrity/FitnessCelebrity.Web/Models/FitnessPathWorkout.cs
@@ -7,10 +7,10 @@ namespace FitnessCelebrity.Web.Models
 {
     public class FitnessPathWorkout
     {
-        public int FitnessPathId { get; set; }
+        public long FitnessPathId { get; set; }
         public FitnessPath FitnessPath { get; set; }
 
-        public int WorkoutId { get; set; }
+        public long WorkoutId { get; set; }
         public Workout Workout { get; set; }
     }
 }
